Move more things into Types to make sure that platform Dnssd knows about them
