+++ src/lib/dnssd/Types.h
@@ -20,6 +20,15 @@
 #include <cstdint>
 #include <cstring>
 
+#include <inet/InetInterface.h>
+#include <lib/core/CHIPError.h>
+#include <lib/core/Optional.h>
+#include <lib/core/PeerId.h>
+#include <lib/dnssd/Constants.h>
+#include <lib/support/BytesToHex.h>
+#include <lib/support/logging/CHIPLogging.h>
+#include <messaging/ReliableMessageProtocolConfig.h>
+
 namespace chip {
 namespace Dnssd {
 
@@ -67,6 +76,245 @@ enum class DiscoveryType
     kOperational,
     kCommissionableNode,
     kCommissionerNode
+};
+
+struct CommonResolutionData
+{
+    static constexpr unsigned kMaxIPAddresses = CHIP_DEVICE_CONFIG_MAX_DISCOVERED_IP_ADDRESSES;
+
+    Inet::InterfaceId interfaceId;
+
+    size_t numIPs = 0; // number of valid IP addresses
+    Inet::IPAddress ipAddress[kMaxIPAddresses];
+
+    uint16_t port                         = 0;
+    char hostName[kHostNameMaxLength + 1] = {};
+    bool supportsTcp                      = false;
+    Optional<bool> isICDOperatingAsLIT;
+    Optional<System::Clock::Milliseconds32> mrpRetryIntervalIdle;
+    Optional<System::Clock::Milliseconds32> mrpRetryIntervalActive;
+    Optional<System::Clock::Milliseconds16> mrpRetryActiveThreshold;
+
+    CommonResolutionData() { Reset(); }
+
+    bool IsValid() const { return !IsHost("") && (numIPs > 0) && (ipAddress[0] != chip::Inet::IPAddress::Any); }
+
+    ReliableMessageProtocolConfig GetRemoteMRPConfig() const
+    {
+        const ReliableMessageProtocolConfig defaultConfig = GetDefaultMRPConfig();
+        return ReliableMessageProtocolConfig(GetMrpRetryIntervalIdle().ValueOr(defaultConfig.mIdleRetransTimeout),
+                                             GetMrpRetryIntervalActive().ValueOr(defaultConfig.mActiveRetransTimeout),
+                                             GetMrpRetryActiveThreshold().ValueOr(defaultConfig.mActiveThresholdTime));
+    }
+    Optional<System::Clock::Milliseconds32> GetMrpRetryIntervalIdle() const { return mrpRetryIntervalIdle; }
+    Optional<System::Clock::Milliseconds32> GetMrpRetryIntervalActive() const { return mrpRetryIntervalActive; }
+    Optional<System::Clock::Milliseconds16> GetMrpRetryActiveThreshold() const { return mrpRetryActiveThreshold; }
+
+    bool IsDeviceTreatedAsSleepy(const ReliableMessageProtocolConfig * defaultMRPConfig) const
+    {
+        // If either sleepy interval (Idle - SII, Active - SAI) has a value and that value is greater
+        // than the value passed to this function, then the peer device will be treated as if it is
+        // a Sleepy End Device (SED)
+        return (mrpRetryIntervalIdle.HasValue() && (mrpRetryIntervalIdle.Value() > defaultMRPConfig->mIdleRetransTimeout)) ||
+            (mrpRetryIntervalActive.HasValue() && (mrpRetryIntervalActive.Value() > defaultMRPConfig->mActiveRetransTimeout));
+    }
+
+    bool IsHost(const char * host) const { return strcmp(host, hostName) == 0; }
+
+    void Reset()
+    {
+        memset(hostName, 0, sizeof(hostName));
+        mrpRetryIntervalIdle    = NullOptional;
+        mrpRetryIntervalActive  = NullOptional;
+        mrpRetryActiveThreshold = NullOptional;
+        isICDOperatingAsLIT     = NullOptional;
+        numIPs                  = 0;
+        port                    = 0;
+        supportsTcp             = false;
+        interfaceId             = Inet::InterfaceId::Null();
+        for (auto & addr : ipAddress)
+        {
+            addr = chip::Inet::IPAddress::Any;
+        }
+    }
+
+    void LogDetail() const
+    {
+        if (!IsHost(""))
+        {
+            ChipLogDetail(Discovery, "\tHostname: %s", hostName);
+        }
+#if CHIP_DETAIL_LOGGING
+        for (unsigned j = 0; j < numIPs; j++)
+        {
+            char buf[Inet::IPAddress::kMaxStringLength];
+            char * ipAddressOut = ipAddress[j].ToString(buf);
+            ChipLogDetail(Discovery, "\tIP Address #%d: %s", j + 1, ipAddressOut);
+        }
+#endif // CHIP_DETAIL_LOGGING
+        if (port > 0)
+        {
+            ChipLogDetail(Discovery, "\tPort: %u", port);
+        }
+        if (mrpRetryIntervalIdle.HasValue())
+        {
+            ChipLogDetail(Discovery, "\tMrp Interval idle: %" PRIu32 " ms", mrpRetryIntervalIdle.Value().count());
+        }
+        else
+        {
+            ChipLogDetail(Discovery, "\tMrp Interval idle: not present");
+        }
+        if (mrpRetryIntervalActive.HasValue())
+        {
+            ChipLogDetail(Discovery, "\tMrp Interval active: %" PRIu32 " ms", mrpRetryIntervalActive.Value().count());
+        }
+        else
+        {
+            ChipLogDetail(Discovery, "\tMrp Interval active: not present");
+        }
+        if (mrpRetryActiveThreshold.HasValue())
+        {
+            ChipLogDetail(Discovery, "\tMrp Active Threshold: %u ms", mrpRetryActiveThreshold.Value().count());
+        }
+        else
+        {
+            ChipLogDetail(Discovery, "\tMrp Active Threshold: not present");
+        }
+        ChipLogDetail(Discovery, "\tTCP Supported: %d", supportsTcp);
+        if (isICDOperatingAsLIT.HasValue())
+        {
+            ChipLogDetail(Discovery, "\tThe ICD operates in %s", isICDOperatingAsLIT.Value() ? "LIT" : "SIT");
+        }
+        else
+        {
+            ChipLogDetail(Discovery, "\tICD: not present");
+        }
+    }
+};
+
+/// Data that is specific to Operational Discovery of nodes
+struct OperationalNodeData
+{
+    PeerId peerId;
+
+    void Reset() { peerId = PeerId(); }
+};
+
+inline constexpr size_t kMaxDeviceNameLen         = 32;
+inline constexpr size_t kMaxRotatingIdLen         = 50;
+inline constexpr size_t kMaxPairingInstructionLen = 128;
+
+/// Data that is specific to commisionable/commissioning node discovery
+struct CommissionNodeData
+{
+    char instanceName[Commission::kInstanceNameMaxLength + 1] = {};
+    uint16_t longDiscriminator                                = 0;
+    uint16_t vendorId                                         = 0;
+    uint16_t productId                                        = 0;
+    uint8_t commissioningMode                                 = 0;
+    uint32_t deviceType                                       = 0;
+    char deviceName[kMaxDeviceNameLen + 1]                    = {};
+    uint8_t rotatingId[kMaxRotatingIdLen]                     = {};
+    size_t rotatingIdLen                                      = 0;
+    uint16_t pairingHint                                      = 0;
+    char pairingInstruction[kMaxPairingInstructionLen + 1]    = {};
+    uint8_t commissionerPasscode                              = 0;
+
+    CommissionNodeData() {}
+
+    void Reset()
+    {
+        // Let constructor clear things as default
+        this->~CommissionNodeData();
+        new (this) CommissionNodeData();
+    }
+
+    bool IsInstanceName(const char * instance) const { return strcmp(instance, instanceName) == 0; }
+
+    void LogDetail() const
+    {
+        if (rotatingIdLen > 0)
+        {
+            char rotatingIdString[chip::Dnssd::kMaxRotatingIdLen * 2 + 1] = "";
+            Encoding::BytesToUppercaseHexString(rotatingId, rotatingIdLen, rotatingIdString, sizeof(rotatingIdString));
+            ChipLogDetail(Discovery, "\tRotating ID: %s", rotatingIdString);
+        }
+        if (strlen(deviceName) != 0)
+        {
+            ChipLogDetail(Discovery, "\tDevice Name: %s", deviceName);
+        }
+        if (vendorId > 0)
+        {
+            ChipLogDetail(Discovery, "\tVendor ID: %u", vendorId);
+        }
+        if (productId > 0)
+        {
+            ChipLogDetail(Discovery, "\tProduct ID: %u", productId);
+        }
+        if (deviceType > 0)
+        {
+            ChipLogDetail(Discovery, "\tDevice Type: %" PRIu32, deviceType);
+        }
+        if (longDiscriminator > 0)
+        {
+            ChipLogDetail(Discovery, "\tLong Discriminator: %u", longDiscriminator);
+        }
+        if (strlen(pairingInstruction) != 0)
+        {
+            ChipLogDetail(Discovery, "\tPairing Instruction: %s", pairingInstruction);
+        }
+        if (pairingHint > 0)
+        {
+            ChipLogDetail(Discovery, "\tPairing Hint: %u", pairingHint);
+        }
+        if (!IsInstanceName(""))
+        {
+            ChipLogDetail(Discovery, "\tInstance Name: %s", instanceName);
+        }
+        ChipLogDetail(Discovery, "\tCommissioning Mode: %u", commissioningMode);
+        if (commissionerPasscode > 0)
+        {
+            ChipLogDetail(Discovery, "\tCommissioner Passcode: %u", commissionerPasscode);
+        }
+    }
+};
+
+struct ResolvedNodeData
+{
+    CommonResolutionData resolutionData;
+    OperationalNodeData operationalData;
+
+    void LogNodeIdResolved() const
+    {
+#if CHIP_PROGRESS_LOGGING
+        // Would be nice to log the interface id, but sorting out how to do so
+        // across our different InterfaceId implementations is a pain.
+        ChipLogProgress(Discovery, "Node ID resolved for " ChipLogFormatX64 ":" ChipLogFormatX64,
+                        ChipLogValueX64(operationalData.peerId.GetCompressedFabricId()),
+                        ChipLogValueX64(operationalData.peerId.GetNodeId()));
+        resolutionData.LogDetail();
+#endif // CHIP_PROGRESS_LOGGING
+    }
+};
+
+struct DiscoveredNodeData
+{
+    CommonResolutionData resolutionData;
+    CommissionNodeData commissionData;
+
+    void Reset()
+    {
+        resolutionData.Reset();
+        commissionData.Reset();
+    }
+    DiscoveredNodeData() { Reset(); }
+
+    void LogDetail() const
+    {
+        ChipLogDetail(Discovery, "Discovered node:");
+        resolutionData.LogDetail();
+        commissionData.LogDetail();
+    }
 };
 
 } // namespace Dnssd
